Consolidate .solcover config files
