--- .solcover.extensions.js
@@ -1,5 +1,0 @@
-const config = require("./.solcover.js")
-
-config.istanbulFolder = "./coverage-extensions"
-
-module.exports = config
