--- .solcover.reputation.js
@@ -1,5 +1,0 @@
-const config = require("./.solcover.js")
-
-config.istanbulFolder = "./coverage-reputation"
-
-module.exports = config
