+++ .solcover/chainid.js
@@ -1,4 +1,4 @@
-const config = require("./.solcover.js")
+const config = require("./solcover.js")
 
 config.providerOptions.network_id = parseInt(process.env.CHAIN_ID, 10);
 config.providerOptions._chainId = parseInt(process.env.CHAIN_ID, 10);
