+++ .solcover/extensions.js
@@ -1,0 +1,5 @@
+const config = require("./solcover.js")
+
+config.istanbulFolder = "./coverage-extensions"
+
+module.exports = config
