+++ .solcover/reputation.js
@@ -1,0 +1,5 @@
+const config = require("./solcover.js")
+
+config.istanbulFolder = "./coverage-reputation"
+
+module.exports = config
